+++ src/problem/func/func_GEN_RED.c
@@ -1,5 +1,5 @@
-/** @file func_GEN_COST.c
- *  @brief This file defines the data structure and routines associated with the function of type GEN_COST.
+/** @file func_GEN_RED.c
+ *  @brief This file defines the data structure and routines associated with the function of type GEN_RED.
  *
  * This file is part of PFNET.
  *
@@ -8,18 +8,18 @@
  * PFNET is released under the BSD 2-clause license.
  */
 
-#include <pfnet/func_GEN_COST.h>
+#include <pfnet/func_GEN_RED.h>
 
-Func* FUNC_GEN_COST_new(REAL weight, Net* net) {
+Func* FUNC_GEN_RED_new(REAL weight, Net* net) {
   Func* f = FUNC_new(weight,net);
-  FUNC_set_func_count_step(f,&FUNC_GEN_COST_count_step);
-  FUNC_set_func_analyze_step(f,&FUNC_GEN_COST_analyze_step);
-  FUNC_set_func_eval_step(f,&FUNC_GEN_COST_eval_step);
-  FUNC_set_name(f,"generation cost");
+  FUNC_set_func_count_step(f,&FUNC_GEN_RED_count_step);
+  FUNC_set_func_analyze_step(f,&FUNC_GEN_RED_analyze_step);
+  FUNC_set_func_eval_step(f,&FUNC_GEN_RED_eval_step);
+  FUNC_set_name(f,"generation redispatch penalty");
   return f;
 }
 
-void FUNC_GEN_COST_count_step(Func* f, Bus* bus, BusDC* busdc, int t) {
+void FUNC_GEN_RED_count_step(Func* f, Bus* bus, BusDC* busdc, int t) {
 
   // Local variables
   Gen* gen;
@@ -34,12 +34,17 @@ void FUNC_GEN_COST_count_step(Func* f, Bus* bus, BusDC* busdc, int t) {
 
   // Generators
   for (gen = BUS_get_gen(bus); gen != NULL; gen = GEN_get_next(gen)) {
-    if (GEN_has_flags(gen,FLAG_VARS,GEN_VAR_P) && !GEN_is_on_outage(gen))
+    
+    // Outage
+    if (GEN_is_on_outage(gen))
+      continue;
+        
+    if (GEN_has_flags(gen,FLAG_VARS,GEN_VAR_P)) // P var
       (*Hphi_nnz)++;
   }
 }
 
-void FUNC_GEN_COST_analyze_step(Func* f, Bus* bus, BusDC* busdc, int t) {
+void FUNC_GEN_RED_analyze_step(Func* f, Bus* bus, BusDC* busdc, int t) {
 
   // Local variables
   Gen* gen;
@@ -56,7 +61,13 @@ void FUNC_GEN_COST_analyze_step(Func* f, Bus* bus, BusDC* busdc, int t) {
 
   // Generators
   for (gen = BUS_get_gen(bus); gen != NULL; gen = GEN_get_next(gen)) {
-    if (GEN_has_flags(gen,FLAG_VARS,GEN_VAR_P) && !GEN_is_on_outage(gen)) {
+    
+    // Outage
+    if (GEN_is_on_outage(gen))
+      continue;
+        
+    if (GEN_has_flags(gen,FLAG_VARS,GEN_VAR_P)) { // P var
+      
       MAT_set_i(Hphi,*Hphi_nnz,GEN_get_index_P(gen,t));
       MAT_set_j(Hphi,*Hphi_nnz,GEN_get_index_P(gen,t));
       (*Hphi_nnz)++;
@@ -64,7 +75,7 @@ void FUNC_GEN_COST_analyze_step(Func* f, Bus* bus, BusDC* busdc, int t) {
   }
 }
 
-void FUNC_GEN_COST_eval_step(Func* f, Bus* bus, BusDC* busdc, int t, Vec* var_values) {
+void FUNC_GEN_RED_eval_step(Func* f, Bus* bus, BusDC* busdc, int t, Vec* var_values) {
 
   // Local variables
   Gen* gen;
@@ -72,11 +83,8 @@ void FUNC_GEN_COST_eval_step(Func* f, Bus* bus, BusDC* busdc, int t, Vec* var_va
   REAL* gphi;
   REAL* Hphi;
   int* Hphi_nnz;
-  int index_P;
   REAL P;
-  REAL Q0;
-  REAL Q1;
-  REAL Q2;
+  REAL P0;
 
   // Constr data
   phi = FUNC_get_phi_ptr(f);
@@ -90,44 +98,31 @@ void FUNC_GEN_COST_eval_step(Func* f, Bus* bus, BusDC* busdc, int t, Vec* var_va
 
   // Generators
   for (gen = BUS_get_gen(bus); gen != NULL; gen = GEN_get_next(gen)) {
-
+    
     // Outage
     if (GEN_is_on_outage(gen))
       continue;
     
-    // Cost coefficients
-    Q0 = GEN_get_cost_coeff_Q0(gen);
-    Q1 = GEN_get_cost_coeff_Q1(gen);
-    Q2 = GEN_get_cost_coeff_Q2(gen);
-    
-    // Variable
-    if (GEN_has_flags(gen,FLAG_VARS,GEN_VAR_P)) {
-	  
-      // Index
-      index_P = GEN_get_index_P(gen,t);
+    if (GEN_has_flags(gen,FLAG_VARS,GEN_VAR_P)) { // P var
       
-      // P
-      P = VEC_get(var_values,index_P);
+      // Value
+      P0 = GEN_get_P(gen,t);
+      P = VEC_get(var_values,GEN_get_index_P(gen,t));
       
       // phi
-      (*phi) += Q0 + Q1*P + Q2*pow(P,2.);
+      (*phi) += 0.5*pow(P-P0,2.);
       
       // gphi
-      gphi[index_P] = Q1 + 2.*Q2*P;
+      gphi[GEN_get_index_P(gen,t)] = P-P0;
       
       // Hphi
-      Hphi[*Hphi_nnz] = 2.*Q2;
+      Hphi[*Hphi_nnz] = 1.;
       (*Hphi_nnz)++;
     }
-    
-    // Constant
     else {
-      
-      // P
-      P = GEN_get_P(gen,t);
-      
+            
       // phi
-      (*phi) += Q0 + Q1*P + Q2*pow(P,2.);
+      (*phi) += 0.;
     }
   }
 }